pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "authlib>=1.6.5",
    "discord-py>=2.6.4",
    "fastapi>=0.119.0",
    "google-cloud-storage>=3.4.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
